* commit 'd34318796c94828c1ef47554ac19a63264be5adf': (61 commits)
  Always turns on PIC, removes setting of cmake_install_prefix
  Deprecates SharedRayList constructor
  Re-enables (mostly) double-precision and fixes the following issues
  Re-adds mpirun to parallel tests
  Misc Bugfixes
  Adds missing file from Utilities/CMakeLists
  Fixes compiler errors in CPU-only mode and updates GridBins constructor to enable construction from std::vector
  Updates CrossSectionUtilities for MCATK's current interface
  * Updates the cmake install process and creates monteray shared library
  Cleans up files for installation
  Removes Ray.t.hh and cleans up some of the .hh files * Moves all inclusion of .t.hh files into a protected ifdef CUDACC block - these should only be called from within MonteRay anyway, probably.
  Fixes minor compilation errors from last commit.
  Cleans up CopyMemory and Geometry/Cartesian unit tests (WIP)
  Updates a CMakeLists that was missing from last commit/push
  Removes MonteRay_SingleValueCopyMemory
  Removes MonteRayCrossSection and HashLookup (old files)
  Adds CrossSection::read and cleans up some CrossSection unit tests
  Removes references to MonteRayMaterial and replace with Material.  Same with MaterialList.  Deletes old MonteRayMaterial files.
  Minor tidying of RayTrace/nightly
  Fixes bug in QuadraticRootFinder (nans in math) and removes testing code in CylindricalGrid
  ...

+++ src/MonteRay/GPU_Utilities/gpuAddTwoDoubles.hh
@@ -5,8 +5,8 @@
 
 namespace MonteRay {
 
-CUDA_CALLABLE_KERNEL  add_double(unsigned N, int *a, int *b, int *c );
-CUDA_CALLABLE_KERNEL  add_single(unsigned N, int *a, int *b, int *c );
+CUDA_CALLABLE_KERNEL  add_double(unsigned N, double *a, double *b, double *c );
+CUDA_CALLABLE_KERNEL  add_single(unsigned N, float *a, float *b, float *c );
 
 double gpuAddTwoDoubles( double, double );
 float  gpuAddTwoFloats(  float, float );
